+++ src/lib.rs
@@ -68,6 +68,8 @@ mod kv;
 mod request;
 mod session;
 
+pub mod payload;
+
 pub use agent::*;
 pub use catalog::*;
 pub use connect_ca::*;
@@ -147,29 +149,6 @@ impl Config {
             }
         })
     }
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct QueryOptions {
-    pub datacenter: Option<String>,
-    pub wait_index: Option<u64>,
-    pub wait_time: Option<Duration>,
-}
-
-#[derive(Clone, Debug)]
-pub struct QueryMeta {
-    pub last_index: Option<u64>,
-    pub request_time: Duration,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct WriteOptions {
-    pub datacenter: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-pub struct WriteMeta {
-    pub request_time: Duration,
 }
 
 pub(crate) mod sealed {